refactor(artifacts): add in memory artifacts repository with default values
