+++ src/core/domain/entities/circlet-artifact.ts
@@ -1,16 +1,8 @@
+import { CircletMainStatType } from '../models/circlet-artifact-data';
 import { MainStat, PossibleMainStats } from '../models/main-statistics';
 import { SetNames } from '../models/sets-with-effects';
 import { PossibleSubStats, SubStatsValues } from '../models/sub-statistics';
 import { Artifact } from './artifact';
-
-export type CircletMainStatType =
-  | PossibleMainStats.percentAtk
-  | PossibleMainStats.percentDef
-  | PossibleMainStats.percentHp
-  | PossibleMainStats.elementalMastery
-  | PossibleMainStats.critRate
-  | PossibleMainStats.cryoDmg
-  | PossibleMainStats.healingBonus;
 
 export class CircletArtifact extends Artifact {
   public mainStat: MainStat;
