+++ src/core/domain/models/artifact-data.ts
@@ -1,0 +1,9 @@
+import { SetNames } from './sets-with-effects';
+import { SubStatsValues } from './sub-statistics';
+
+export interface ArtifactData {
+  id: string;
+  set?: SetNames;
+  level?: number;
+  subStats: SubStatsValues;
+}
