+++ src/core/usescases/build-optimizer.spec.ts
@@ -1,15 +1,20 @@
 import { PossibleMainStats } from '../domain/models/main-statistics';
-import { PossibleSubStats, SubStatsValues } from '../domain/models/sub-statistics';
+import { PossibleSubStats } from '../domain/models/sub-statistics';
 import { BuildOptimizer } from './build-optimizer';
 import { PossibleCharacterStats } from '../domain/models/character-statistics';
 import { SetNames } from '../domain/models/sets-with-effects';
 import { FlowerArtifact } from '../domain/entities/flower-artifact';
 import { PlumeArtifact } from '../domain/entities/plume-artifact';
-import { CircletArtifact, CircletMainStatType } from '../domain/entities/circlet-artifact';
-import { GobletArtifact, GobletMainStatType } from '../domain/entities/goblet-artifact';
-import { SandsArtifact, SandsMainStatType } from '../domain/entities/sands-artifact';
+import { CircletArtifact } from '../domain/entities/circlet-artifact';
+import { GobletArtifact } from '../domain/entities/goblet-artifact';
+import { SandsArtifact } from '../domain/entities/sands-artifact';
 import { InMemoryCharactersRepository } from '../adapters/secondaries/in-mermory-characters-repository';
 import { Character } from '../domain/models/character';
+import { ArtifactData } from '../domain/models/artifact-data';
+import { CircletArtifactData } from '../domain/models/circlet-artifact-data';
+import { GobletArtifactData } from '../domain/models/goblet-artifact-data';
+import { SandsArtifactData } from '../domain/models/sands-artifact-data';
+import { InMemoryArtifactsRepository } from '../adapters/secondaries/in-mermory-artifacts-repository';
 
 describe('BuildOptimizer.computeBuildStats', () => {
   let buildOptimizer: BuildOptimizer;
@@ -1049,185 +1054,13 @@ describe('BuildOptimizer.computeBuildStats', () => {
 
 describe('BuildOptimizer.filterArtifacts', () => {
   let buildOptimizer: BuildOptimizer;
+  const artifactsRepository: InMemoryArtifactsRepository = new InMemoryArtifactsRepository();
 
-  const allFlowerArtifacts: FlowerArtifact[] = getFlowerArtifactsWithValues([
-    {
-      id: '0',
-      set: SetNames.lavawalker,
-      level: 2,
-      subStats: {
-        [PossibleSubStats.flatAtk]: 5,
-        [PossibleSubStats.critRate]: 3.2,
-        [PossibleSubStats.percentAtk]: 3,
-        [PossibleSubStats.critDmg]: 3.2,
-      },
-    },
-    {
-      id: '1',
-      set: SetNames.thunderingFury,
-      level: 8,
-      subStats: {
-        [PossibleSubStats.energyRecharge]: 3,
-        [PossibleSubStats.percentHp]: 6,
-        [PossibleSubStats.critDmg]: 3.9,
-        [PossibleSubStats.percentAtk]: 7,
-      },
-    },
-  ]);
-
-  const allPlumeArtifacts: PlumeArtifact[] = getPlumeArtifactsWithValues([
-    {
-      id: '2',
-      set: SetNames.retracingBolide,
-      level: 7,
-      subStats: {
-        [PossibleSubStats.energyRecharge]: 4,
-        [PossibleSubStats.flatDef]: 7,
-        [PossibleSubStats.critRate]: 2.7,
-        [PossibleSubStats.critDmg]: 5,
-      },
-    },
-    {
-      id: '3',
-      set: SetNames.blizzardStrayer,
-      level: 12,
-      subStats: {
-        [PossibleSubStats.percentAtk]: 5,
-        [PossibleSubStats.flatHp]: 12,
-        [PossibleSubStats.flatDef]: 6,
-        [PossibleSubStats.percentDef]: 8,
-      },
-    },
-  ]);
-
-  const allSandsArtifacts: SandsArtifact[] = getSandsArtifactsWithValues([
-    {
-      id: '4',
-      set: SetNames.thunderingFury,
-      level: 12,
-      mainStatType: PossibleMainStats.percentHp,
-      subStats: {
-        [PossibleSubStats.percentDef]: 6,
-        [PossibleSubStats.elementalMastery]: 7,
-        [PossibleSubStats.critRate]: 3.2,
-        [PossibleSubStats.critDmg]: 2.9,
-      },
-    },
-    {
-      id: '5',
-      set: SetNames.bloodstainedChivalry,
-      level: 16,
-      mainStatType: PossibleMainStats.percentAtk,
-      subStats: {
-        [PossibleSubStats.energyRecharge]: 6,
-        [PossibleSubStats.flatAtk]: 7,
-        [PossibleSubStats.critRate]: 3.2,
-        [PossibleSubStats.flatDef]: 2.9,
-      },
-    },
-    {
-      id: '6',
-      set: SetNames.retracingBolide,
-      level: 8,
-      mainStatType: PossibleMainStats.elementalMastery,
-      subStats: {
-        [PossibleSubStats.flatHp]: 6,
-        [PossibleSubStats.critDmg]: 7,
-        [PossibleSubStats.percentDef]: 3.2,
-        [PossibleSubStats.percentAtk]: 2.9,
-      },
-    },
-  ]);
-
-  const allGobletArtifacts: GobletArtifact[] = getGobletArtifactsWithValues([
-    {
-      id: '5',
-      set: SetNames.lavawalker,
-      level: 15,
-      mainStatType: PossibleMainStats.percentDef,
-      subStats: {
-        [PossibleSubStats.critRate]: 2.5,
-        [PossibleSubStats.percentHp]: 5.2,
-        [PossibleSubStats.percentAtk]: 4,
-        [PossibleSubStats.flatHp]: 3,
-      },
-    },
-    {
-      id: '5',
-      set: SetNames.archaicPetra,
-      level: 12,
-      mainStatType: PossibleMainStats.percentDef,
-      subStats: {
-        [PossibleSubStats.critDmg]: 2.5,
-        [PossibleSubStats.energyRecharge]: 5.2,
-        [PossibleSubStats.percentHp]: 4,
-        [PossibleSubStats.flatDef]: 3,
-      },
-    },
-    {
-      id: '6',
-      set: SetNames.blizzardStrayer,
-      level: 8,
-      mainStatType: PossibleMainStats.cryoDmg,
-      subStats: {
-        [PossibleSubStats.elementalMastery]: 4,
-        [PossibleSubStats.percentHp]: 5.2,
-        [PossibleSubStats.percentAtk]: 4,
-        [PossibleSubStats.critDmg]: 3,
-      },
-    },
-    {
-      id: '6',
-      set: SetNames.wanderersTroupe,
-      level: 4,
-      mainStatType: PossibleMainStats.physicalDmg,
-      subStats: {
-        [PossibleSubStats.flatAtk]: 4,
-        [PossibleSubStats.percentDef]: 5.2,
-        [PossibleSubStats.energyRecharge]: 4,
-        [PossibleSubStats.critRate]: 3,
-      },
-    },
-  ]);
-
-  const allCircletArtifacts: CircletArtifact[] = getCircletArtifactsWithValues([
-    {
-      id: '7',
-      set: SetNames.retracingBolide,
-      level: 17,
-      mainStatType: PossibleMainStats.healingBonus,
-      subStats: {
-        [PossibleSubStats.percentDef]: 4,
-        [PossibleSubStats.flatAtk]: 4,
-        [PossibleSubStats.critDmg]: 3.2,
-        [PossibleSubStats.percentHp]: 5,
-      },
-    },
-    {
-      id: '7',
-      set: SetNames.gladiatorsFinale,
-      level: 12,
-      mainStatType: PossibleMainStats.critRate,
-      subStats: {
-        [PossibleSubStats.flatHp]: 4,
-        [PossibleSubStats.percentAtk]: 4,
-        [PossibleSubStats.critDmg]: 3.2,
-        [PossibleSubStats.energyRecharge]: 5,
-      },
-    },
-    {
-      id: '7',
-      set: SetNames.retracingBolide,
-      level: 14,
-      mainStatType: PossibleMainStats.critRate,
-      subStats: {
-        [PossibleSubStats.energyRecharge]: 4,
-        [PossibleSubStats.flatAtk]: 4,
-        [PossibleSubStats.critDmg]: 3.2,
-        [PossibleSubStats.flatHp]: 5,
-      },
-    },
-  ]);
+  const allFlowerArtifacts: FlowerArtifact[] = artifactsRepository.getFlowerArtifacts();
+  const allPlumeArtifacts: PlumeArtifact[] = artifactsRepository.getPlumeArtifacts();
+  const allSandsArtifacts: SandsArtifact[] = artifactsRepository.getSandsArtifacts();
+  const allGobletArtifacts: GobletArtifact[] = artifactsRepository.getGobletArtifacts();
+  const allCircletArtifacts: CircletArtifact[] = artifactsRepository.getCircletArtifacts();
 
   beforeEach(() => {
     buildOptimizer = new BuildOptimizer(allFlowerArtifacts, allPlumeArtifacts, allSandsArtifacts, allGobletArtifacts, allCircletArtifacts);
@@ -1299,41 +1132,19 @@ function getLevel(artifactLevel: number): number {
   return artifactLevel ? artifactLevel : 0;
 }
 
-function getFlowerArtifactsWithValues(
-  allArtifactsData: {
-    id: string;
-    set?: SetNames;
-    level?: number;
-    subStats: SubStatsValues;
-  }[],
-): FlowerArtifact[] {
+function getFlowerArtifactsWithValues(allArtifactsData: ArtifactData[]): FlowerArtifact[] {
   return allArtifactsData.map(
     (artifactData) => new FlowerArtifact(artifactData.id, getSet(artifactData.set), artifactData.subStats, getLevel(artifactData.level)),
   );
 }
 
-function getPlumeArtifactsWithValues(
-  allArtifactsData: {
-    id: string;
-    set?: SetNames;
-    level?: number;
-    subStats: SubStatsValues;
-  }[],
-): PlumeArtifact[] {
+function getPlumeArtifactsWithValues(allArtifactsData: ArtifactData[]): PlumeArtifact[] {
   return allArtifactsData.map(
     (artifactData) => new PlumeArtifact(artifactData.id, getSet(artifactData.set), artifactData.subStats, getLevel(artifactData.level)),
   );
 }
 
-function getSandsArtifactsWithValues(
-  allArtifactsData: {
-    id: string;
-    set?: SetNames;
-    level?: number;
-    subStats: SubStatsValues;
-    mainStatType: SandsMainStatType;
-  }[],
-): SandsArtifact[] {
+function getSandsArtifactsWithValues(allArtifactsData: SandsArtifactData[]): SandsArtifact[] {
   return allArtifactsData.map(
     (artifactData) =>
       new SandsArtifact(
@@ -1346,15 +1157,7 @@ function getSandsArtifactsWithValues(
   );
 }
 
-function getGobletArtifactsWithValues(
-  allArtifactsData: {
-    id: string;
-    set?: SetNames;
-    level?: number;
-    subStats: SubStatsValues;
-    mainStatType: GobletMainStatType;
-  }[],
-): GobletArtifact[] {
+function getGobletArtifactsWithValues(allArtifactsData: GobletArtifactData[]): GobletArtifact[] {
   return allArtifactsData.map(
     (artifactData) =>
       new GobletArtifact(
@@ -1367,15 +1170,7 @@ function getGobletArtifactsWithValues(
   );
 }
 
-function getCircletArtifactsWithValues(
-  allArtifactsData: {
-    id: string;
-    set?: SetNames;
-    level?: number;
-    subStats: SubStatsValues;
-    mainStatType: CircletMainStatType;
-  }[],
-): CircletArtifact[] {
+function getCircletArtifactsWithValues(allArtifactsData: CircletArtifactData[]): CircletArtifact[] {
   return allArtifactsData.map(
     (artifactData) =>
       new CircletArtifact(
